chore(mdx): move custom mdx component to dedicated folders to increase overview
